fix: 내정보 페이지 스타일 및 form 수정
