+++ src/components/type/userType.ts
@@ -1,0 +1,13 @@
+interface TechStack {
+  value?: number;
+  id?: number;
+  name?: string;
+  label?: string;
+}
+
+export interface IuserType {
+  id: string;
+  techStackDtos: TechStack[];
+  profileImage: string;
+  nickname: string;
+}
